parser/pyproject.toml
[tool.poetry]
name = "w24ok-bot-parser"
version = "0.1.0"
description = ""
authors = ["ArthurGolubev <gg.arthur.go@gmail.com>"]
readme = "README.md"
packages = [{include = "w24ok_bot_parser"}]

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
selenium = "^4.8.2"
notifiers = "^1.3.3"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
